Refactor: move Fd and File to core
