+++ src/AtomicFile.hpp
@@ -20,8 +20,6 @@
 
 #include "core/system.hpp"
 
-#include "config.h"
-
 #include <string>
 #include <vector>
