+++ src/Hash.cpp
@@ -18,7 +18,7 @@
 
 #include "Hash.hpp"
 
-#include "Fd.hpp"
+#include "core/Fd.hpp"
 #include "Logging.hpp"
 #include "core/fmtmacros.hpp"
