+++ src/Logging.cpp
@@ -20,11 +20,10 @@
 #include "Logging.hpp"
 
 #include "Config.hpp"
-#include "File.hpp"
+#include "core/File.hpp"
 #include "Util.hpp"
 #include "core/fmtmacros.hpp"
 #include "exceptions.hpp"
-#include "execute.hpp"
 
 #ifdef HAVE_SYSLOG_H
 #  include <syslog.h>
