+++ src/Result.cpp
@@ -23,8 +23,8 @@
 #include "CacheEntryWriter.hpp"
 #include "Config.hpp"
 #include "Context.hpp"
-#include "Fd.hpp"
-#include "File.hpp"
+#include "core/Fd.hpp"
+#include "core/File.hpp"
 #include "Logging.hpp"
 #include "Stat.hpp"
 #include "Statistics.hpp"
