+++ src/ResultExtractor.hpp
@@ -20,7 +20,7 @@
 
 #include "core/system.hpp"
 
-#include "Fd.hpp"
+#include "core/Fd.hpp"
 #include "Result.hpp"
 
 class Context;
