+++ src/Util.cpp
@@ -20,7 +20,7 @@
 
 #include "Config.hpp"
 #include "Context.hpp"
-#include "Fd.hpp"
+#include "core/Fd.hpp"
 #include "Logging.hpp"
 #include "TemporaryFile.hpp"
 #include "core/FormatNonstdStringView.hpp"
