+++ src/ccache.cpp
@@ -25,8 +25,8 @@
 #include "Compression.hpp"
 #include "Context.hpp"
 #include "Depfile.hpp"
-#include "Fd.hpp"
-#include "File.hpp"
+#include "core/Fd.hpp"
+#include "core/File.hpp"
 #include "Finalizer.hpp"
 #include "Hash.hpp"
 #include "Lockfile.hpp"
