+++ unittest/test_Util.cpp
@@ -17,9 +17,9 @@
 // Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
 #include "../src/Config.hpp"
-#include "../src/Fd.hpp"
 #include "../src/Util.hpp"
 #include "TestUtil.hpp"
+#include "core/Fd.hpp"
 #include "core/fmtmacros.hpp"
 
 #include "third_party/doctest.h"
